requirements.txt
pandas
numpy
scipy
yfinance
pyyaml
streamlit
plotly
pytest
numpy-financial
